Call postcode validation before adding/deleting postcodes

+++ src/main/java/uk/gov/hmcts/dts/fact/controllers/admin/AdminCourtPostcodeController.java
@@ -5,9 +5,12 @@ import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
 import uk.gov.hmcts.dts.fact.config.security.Role;
+import uk.gov.hmcts.dts.fact.exception.InvalidPostcodeException;
 import uk.gov.hmcts.dts.fact.services.admin.AdminCourtPostcodeService;
+import uk.gov.hmcts.dts.fact.services.validation.ValidationService;
 
 import java.net.URI;
 import java.util.List;
@@ -24,10 +27,12 @@ import static uk.gov.hmcts.dts.fact.services.admin.AdminRole.FACT_SUPER_ADMIN;
 )
 public class AdminCourtPostcodeController {
     private final AdminCourtPostcodeService adminService;
+    private final ValidationService validationService;
 
     @Autowired
-    public AdminCourtPostcodeController(AdminCourtPostcodeService adminService) {
+    public AdminCourtPostcodeController(AdminCourtPostcodeService adminService, ValidationService validationService) {
         this.adminService = adminService;
+        this.validationService = validationService;
     }
 
     @GetMapping(path = "/{slug}/postcodes")
@@ -41,15 +46,23 @@ public class AdminCourtPostcodeController {
     @ApiOperation("Add new court postcodes")
     @Role(FACT_SUPER_ADMIN)
     public ResponseEntity<List<String>> addCourtPostcodes(@PathVariable String slug, @RequestBody List<String> postcodes) {
-        final List<String> postcodeCreated = adminService.addCourtPostcodes(slug, postcodes);
-        return created(URI.create(StringUtils.EMPTY)).body(postcodeCreated);
+        final List<String> invalidPostcodes = validationService.validatePostcodes(postcodes);
+        if (CollectionUtils.isEmpty(invalidPostcodes)) {
+            final List<String> postcodeCreated = adminService.addCourtPostcodes(slug, postcodes);
+            return created(URI.create(StringUtils.EMPTY)).body(postcodeCreated);
+        }
+        throw new InvalidPostcodeException(invalidPostcodes);
     }
 
     @DeleteMapping(path = "/{slug}/postcodes")
     @ApiOperation("Delete court postcodes")
     @Role(FACT_SUPER_ADMIN)
     public ResponseEntity deleteCourtPostcodes(@PathVariable String slug, @RequestBody List<String> postcodes) {
-        adminService.deleteCourtPostcodes(slug, postcodes);
-        return ok().build();
+        final List<String> invalidPostcodes = validationService.validatePostcodes(postcodes);
+        if (CollectionUtils.isEmpty(invalidPostcodes)) {
+            adminService.deleteCourtPostcodes(slug, postcodes);
+            return ok().build();
+        }
+        throw new InvalidPostcodeException(invalidPostcodes);
     }
 }
